Fixed bug
Created a BookSearchView to separate book searching from the content view. Added it to the navigation and fixed the bug that was blocking it from displaying.
