+++ PageTurnerBooks/ViewModels/BookManager.swift
@@ -9,6 +9,7 @@ import Foundation
 import SwiftUI
 
 class BookManager: CoordinatorDelegate, ObservableObject {
+    static let shared = BookManager()
     @Published var books: [BookItem] = []
 
     func didRetrieveBooks(_ bookItems: [BookItem]) {
